+++ avaje-jex/src/main/java/io/avaje/jex/Jex.java
@@ -10,7 +10,6 @@ import io.avaje.jex.http.ExceptionHandler;
 import io.avaje.jex.http.ExchangeHandler;
 import io.avaje.jex.http.HttpFilter;
 import io.avaje.jex.http.sse.SseClient;
-import io.avaje.jex.http.sse.SseHandler;
 import io.avaje.jex.security.Role;
 import io.avaje.jex.spi.JexPlugin;
 import io.avaje.jex.spi.JsonService;
